Splitting into translation units
